Refactor: Mapper Class 삭제, dto 패키지 경로 변경 및 OrderStatus.DONE 인 주문만 조회하여 정산 (#94)
