--- src/main/java/com/bttf/queosk/mapper/SettlementMapper.java
@@ -1,13 +1,0 @@
-package com.bttf.queosk.mapper;
-
-import com.bttf.queosk.dto.settlementdto.SettlementDto;
-import com.bttf.queosk.dto.settlementdto.SettlementResponse;
-import org.mapstruct.Mapper;
-import org.mapstruct.factory.Mappers;
-
-@Mapper
-public interface SettlementMapper {
-    SettlementMapper INSTANCE = Mappers.getMapper(SettlementMapper.class);
-
-    SettlementResponse settlementDtoToSettlementResponse(SettlementDto settlementDto);
-}
