+++ src/main/java/com/bttf/queosk/repository/SettlementRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.bttf.queosk.repository;
 
-import com.bttf.queosk.dto.settlementdto.SettlementDto;
-import com.bttf.queosk.entity.QMenu;
+import com.bttf.queosk.dto.SettlementDto;
 import com.bttf.queosk.entity.QOrder;
+import com.bttf.queosk.enumerate.OrderStatus;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +18,6 @@ public class SettlementRepositoryImpl implements SettlementRepository {
 
     public List<SettlementDto.OrderdMenu> getTodaySales(Long restaurantId) {
         QOrder order = QOrder.order;
-        QMenu menu = QMenu.menu;
 
         LocalDate today = LocalDate.now();
         LocalDateTime startDateTime = LocalDateTime.of(today, LocalTime.MIN);
@@ -30,22 +29,21 @@ public class SettlementRepositoryImpl implements SettlementRepository {
                 .select(Projections.constructor(SettlementDto.OrderdMenu.class, order.menu.name, order.count, order.menu.price))
                 .from(order)
                 .where(order.restaurant.id.eq(restaurantId)
-                        .and(order.createdAt.between(startDateTime, endDateTime)))
+                        .and(order.createdAt.between(startDateTime, endDateTime))
+                        .and(order.status.eq(OrderStatus.DONE)))
                 .groupBy(order.menu)
                 .fetch();
     }
 
     public List<SettlementDto.OrderdMenu> getPeriodSales(Long restaurantId, LocalDateTime to, LocalDateTime from) {
         QOrder order = QOrder.order;
-        QMenu menu = QMenu.menu;
-
-        LocalDate today = LocalDate.now();
 
         return jpaQueryFactory
                 .select(Projections.constructor(SettlementDto.OrderdMenu.class, order.menu.name, order.count, order.menu.price))
                 .from(order)
                 .where(order.restaurant.id.eq(restaurantId)
-                        .and(order.createdAt.between(from, to)))
+                        .and(order.createdAt.between(from, to))
+                        .and(order.status.eq(OrderStatus.DONE)))
                 .groupBy(order.menu)
                 .fetch();
     }
